# chore: faster Perps format functions + transformMarketData